All script references fixed
