+++ Assets/Scripts/Terrain/DetailTextureList.cs.meta
