--- Assets/Scripts/Terrain/Mesh.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b874272b121dfce41bb629d220b54d94
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
