--- Assets/Scripts/Terrain/Mesh/MeshMapInspector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ae0f028d6a11b0746aca0d4da69a356d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
