+++ Assets/Scripts/Terrain/TerrainManager.cs
@@ -5,35 +5,33 @@ using UnityEngine;
 
 public class TerrainManager : StaticMonoBehaviour<TerrainManager>
 {
-    public MeshMapEditor meshMapEditor;
-    public TerrainInspector terrainInspector;
+    public TerrainEditor terrainEditor;
 
     void Start()
     {
-        meshMapEditor.Enable();
-        terrainInspector.SetTerrainMode();
+        terrainEditor.Enable();
     }
 
     public void PopulateCampaign(Campaign campaign)
     {
-        meshMapEditor.SaveIntoCampaign(campaign);
+        terrainEditor.SaveIntoCampaign(campaign);
     }
 
     public void LoadCampaign(Campaign campaign)
     {
         if (campaign.terrainData != null && campaign.terrainData.heightMap != null)
         {
-            meshMapEditor.LoadFromCampaign(campaign);
+            terrainEditor.LoadFromCampaign(campaign);
         }
     }
 
     public bool TerrainNeedsSaved()
     {
-        return meshMapEditor.IsDirty();
+        return terrainEditor.IsDirty();
     }
 
     public void ClearAllTerrain()
     {
-        meshMapEditor.Clear();
+        terrainEditor.Clear();
     }
 }
